checking authorization and authentication on views

+++ inventory/views/core.py
@@ -1,4 +1,4 @@
-from django.http import HttpResponse
+from django.http import HttpResponse, HttpResponseForbidden
 from django.shortcuts import get_object_or_404, redirect, render
 
 from inventory import models
@@ -18,40 +18,51 @@ def browse(request, category_name):
     })
     
 def item_new(request):
-    if request.method == "POST":
-        form = forms.ItemForm(request.POST)
-        if form.is_valid():
-            item = form.save(commit=False)
-            item.author_id = request.user.id
-            item.save()
-            
-            return redirect('home')
+    if not request.user.is_authenticated:
+        return redirect('login')
     else:
-        form = forms.ItemForm()
-    
-    categories = models.Category.objects.all()
-    return render(request, 'item_new.html', {
-        'categories': categories,
-        'form': form
-    })
+        if request.method == "POST":
+            form = forms.ItemForm(request.POST)
+            if form.is_valid():
+                item = form.save(commit=False)
+                item.author_id = request.user.id
+                item.save()
+                
+                return redirect('home')
+        else:
+            form = forms.ItemForm()
+        
+        categories = models.Category.objects.all()
+        return render(request, 'item_new.html', {
+            'categories': categories,
+            'form': form
+        })
     
 def item_edit(request, item_id):
     item = get_object_or_404(models.Item, id=item_id)
-    if request.method == "POST":
-        form = forms.ItemForm(request.POST, instance=item)
-        if form.is_valid():
-            form.save()
-            return redirect('browse', 'Auto')
+    
+    if item.author_id != request.user.id:
+        return HttpResponseForbidden()
     else:
-        form = forms.ItemForm(instance=item)
-        
-    categories = models.Category.objects.all()
-    return render(request, 'item_edit.html', {
-        'categories': categories,
-        'form': form
-    })
+        if request.method == "POST":
+            form = forms.ItemForm(request.POST, instance=item)
+            if form.is_valid():
+                form.save()
+                return redirect('browse', 'Auto')
+        else:
+            form = forms.ItemForm(instance=item)
+            
+        categories = models.Category.objects.all()
+        return render(request, 'item_edit.html', {
+            'categories': categories,
+            'form': form
+        })
     
 def item_delete(request, item_id):
     item = get_object_or_404(models.Item, id=item_id)
-    item.delete()
-    return redirect('browse', item.category.name)
+    
+    if item.author_id != request.user.id:
+        return HttpResponseForbidden()
+    else:
+        item.delete()
+        return redirect('browse', item.category.name)
